Autotiled map feature added (#46)
The map tiles are automatically generated based on a tilesheet texture. 

The current configuration handles two types of terrain: 
Dirst for monster path and Grass otherwise. 

Once the map is loaded, the path will be defined and the tiles drew accordingly. 
Corners, diagonals and curves are properly handled within the 16 tiles present on \assets\sprites\tilesheet.png.